Done some more Dagger modules refactoring
