+++ app/src/main/java/com/aconno/sensorics/dagger/actionlist/ActionListActivityModule.kt
@@ -3,4 +3,4 @@ package com.aconno.sensorics.dagger.actionlist
 import dagger.Module
 
 @Module
-class ActionListModule
+class ActionListActivityModule
