+++ app/src/main/java/com/aconno/sensorics/dagger/application/AppModule.kt
@@ -8,16 +8,8 @@ import com.aconno.sensorics.AlarmServiceControllerImpl
 import com.aconno.sensorics.IntentProviderImpl
 import com.aconno.sensorics.SensoricsApplication
 import com.aconno.sensorics.dagger.time.TimeModule
-import com.aconno.sensorics.data.mapper.*
 import com.aconno.sensorics.data.repository.InMemoryRepositoryImpl
 import com.aconno.sensorics.data.repository.SensoricsDatabase
-import com.aconno.sensorics.data.repository.action.ActionsRepositoryImpl
-import com.aconno.sensorics.data.repository.devices.DeviceMapper
-import com.aconno.sensorics.data.repository.devices.DeviceRepositoryImpl
-import com.aconno.sensorics.data.repository.googlepublish.GooglePublishRepositoryImpl
-import com.aconno.sensorics.data.repository.mqttpublish.MqttPublishRepositoryImpl
-import com.aconno.sensorics.data.repository.publishdevicejoin.PublishDeviceJoinRepositoryImpl
-import com.aconno.sensorics.data.repository.restpublish.RestPublishRepositoryImpl
 import com.aconno.sensorics.data.repository.sync.SyncDao
 import com.aconno.sensorics.data.repository.sync.SyncRepositoryImpl
 import com.aconno.sensorics.device.DeviceAudioManagerImpl
@@ -27,11 +19,9 @@ import com.aconno.sensorics.device.VibratorImpl
 import com.aconno.sensorics.device.notification.IntentProvider
 import com.aconno.sensorics.device.notification.NotificationDisplayImpl
 import com.aconno.sensorics.device.notification.NotificationFactory
-import com.aconno.sensorics.device.time.TimeProviderImpl
 import com.aconno.sensorics.domain.AlarmServiceController
 import com.aconno.sensorics.domain.DeviceAudioManager
 import com.aconno.sensorics.domain.Vibrator
-import com.aconno.sensorics.domain.actions.ActionsRepository
 import com.aconno.sensorics.domain.format.ConnectionCharacteristicsFinder
 import com.aconno.sensorics.domain.format.ConnectionCharacteristicsFinderImpl
 import com.aconno.sensorics.domain.format.FormatMatcher
@@ -41,22 +31,16 @@ import com.aconno.sensorics.domain.interactor.consolidation.GenerateScanDeviceUs
 import com.aconno.sensorics.domain.interactor.convert.ReadingToInputUseCase
 import com.aconno.sensorics.domain.interactor.filter.FilterByFormatUseCase
 import com.aconno.sensorics.domain.interactor.filter.FilterByMacUseCase
-import com.aconno.sensorics.domain.interactor.ifttt.UpdatePublishUseCase
-import com.aconno.sensorics.domain.interactor.repository.*
 import com.aconno.sensorics.domain.interactor.resources.GetFormatsUseCase
-import com.aconno.sensorics.domain.model.Device
-import com.aconno.sensorics.domain.repository.DeviceRepository
 import com.aconno.sensorics.domain.repository.InMemoryRepository
 import com.aconno.sensorics.domain.repository.SyncRepository
 import com.aconno.sensorics.domain.serialization.Deserializer
 import com.aconno.sensorics.domain.serialization.DeserializerImpl
 import com.aconno.sensorics.domain.telephony.DeviceTelephonyManager
-import com.aconno.sensorics.domain.time.TimeProvider
 import com.google.gson.FieldNamingPolicy
 import com.google.gson.GsonBuilder
 import dagger.Module
 import dagger.Provides
-import io.reactivex.Flowable
 import javax.inject.Singleton
 
 @Module (includes = [TimeModule::class])
@@ -109,58 +93,6 @@ class AppModule {
 
     @Provides
     @Singleton
-    fun provideActionsRepository(
-        sensoricsDatabase: SensoricsDatabase
-    ): ActionsRepository {
-        return ActionsRepositoryImpl(sensoricsDatabase.actionDao())
-    }
-
-    @Provides
-    @Singleton
-    fun provideGooglePublishRepository(
-        sensoricsDatabase: SensoricsDatabase,
-        googlePublishEntityDataMapper: GooglePublishEntityDataMapper,
-        googlePublishDataMapper: GooglePublishDataMapper
-    ): GooglePublishRepository {
-        return GooglePublishRepositoryImpl(
-            sensoricsDatabase.googlePublishDao(),
-            googlePublishEntityDataMapper,
-            googlePublishDataMapper
-        )
-    }
-
-    @Provides
-    @Singleton
-    fun provideRESTPublishRepository(
-        sensoricsDatabase: SensoricsDatabase,
-        restPublishEntityDataMapper: RestPublishEntityDataMapper,
-        restPublishDataMapper: RestPublishDataMapper,
-        restHeaderDataMapper: RestHeaderDataMapper,
-        restHttpGetParamDataMapper: RestHttpGetParamDataMapper
-    ): RestPublishRepository {
-        return RestPublishRepositoryImpl(
-            sensoricsDatabase.restPublishDao(),
-            restPublishEntityDataMapper,
-            restPublishDataMapper,
-            restHeaderDataMapper,
-            restHttpGetParamDataMapper
-        )
-    }
-
-    @Provides
-    @Singleton
-    fun provideMqttPublishRepository(
-        sensoricsDatabase: SensoricsDatabase,
-        mqttPublishDataMapper: MqttPublishDataMapper
-    ): MqttPublishRepository {
-        return MqttPublishRepositoryImpl(
-            sensoricsDatabase.mqttPublishDao(),
-            mqttPublishDataMapper
-        )
-    }
-
-    @Provides
-    @Singleton
     fun provideSensoricsDatabase(sensoricsApplication: SensoricsApplication): SensoricsDatabase {
         return Room.databaseBuilder(
             sensoricsApplication,
@@ -192,46 +124,7 @@ class AppModule {
 
     @Provides
     @Singleton
-    fun provideDeviceRepository(
-        sensoricsDatabase: SensoricsDatabase,
-        deviceMapper: DeviceMapper
-    ): DeviceRepository {
-        return DeviceRepositoryImpl(sensoricsDatabase.deviceDao(), deviceMapper)
-    }
-
-    @Provides
-    @Singleton
-    fun provideGetSavedDevicesList(
-        deviceRepository: DeviceRepository
-    ): Flowable<List<Device>> {
-        return GetSavedDevicesUseCase(deviceRepository).execute()
-    }
-
-    @Provides
-    @Singleton
-    fun provideGetSavedDevicesMaybeUseCase(
-        deviceRepository: DeviceRepository
-    ): GetSavedDevicesMaybeUseCase {
-        return GetSavedDevicesMaybeUseCase(deviceRepository)
-    }
-
-    @Provides
-    @Singleton
     fun provideReadingToInputUseCase() = ReadingToInputUseCase()
-
-    @Provides
-    @Singleton
-    fun providePublishDeviceJoinRepository(
-        sensoricsDatabase: SensoricsDatabase,
-        deviceMapper: DeviceMapper,
-        publishDeviceJoinMapper: PublishDeviceJoinMapper
-    ): PublishDeviceJoinRepository {
-        return PublishDeviceJoinRepositoryImpl(
-            sensoricsDatabase.publishDeviceJoinDao(),
-            deviceMapper,
-            publishDeviceJoinMapper
-        )
-    }
 
     @Provides
     @Singleton
@@ -266,39 +159,6 @@ class AppModule {
         GsonBuilder()
             .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
             .create()
-
-    @Provides
-    @Singleton
-    fun provideGetDevicesThatConnectedWithGooglePublishUseCase(
-        publishDeviceJoinRepository: PublishDeviceJoinRepository
-    ): GetDevicesThatConnectedWithGooglePublishUseCase {
-        return GetDevicesThatConnectedWithGooglePublishUseCase(publishDeviceJoinRepository)
-    }
-
-    @Provides
-    @Singleton
-    fun provideGetDevicesThatConnectedWithRESTPublishUseCase(
-        publishDeviceJoinRepository: PublishDeviceJoinRepository
-    ): GetDevicesThatConnectedWithRestPublishUseCase {
-        return GetDevicesThatConnectedWithRestPublishUseCase(publishDeviceJoinRepository)
-    }
-
-    @Provides
-    @Singleton
-    fun provideGetDevicesThatConnectedWithMqttPublishUseCase(
-        publishDeviceJoinRepository: PublishDeviceJoinRepository
-    ): GetDevicesThatConnectedWithMqttPublishUseCase {
-        return GetDevicesThatConnectedWithMqttPublishUseCase(publishDeviceJoinRepository)
-    }
-
-    @Provides
-    @Singleton
-    fun provideUpdatePublishUseCase(
-        googlePublishRepository: GooglePublishRepository,
-        mqttPublishRepository: MqttPublishRepository,
-        restPublishRepository: RestPublishRepository
-    ): UpdatePublishUseCase =
-        UpdatePublishUseCase(googlePublishRepository, mqttPublishRepository, restPublishRepository)
 
     @Provides
     @Singleton
