+++ app/src/main/java/com/aconno/sensorics/dagger/application/ResourcesModule.kt
@@ -117,4 +117,6 @@ class ResourcesModule {
             application.cacheDir.absolutePath
         )
     }
+
+
 }
