+++ app/src/main/java/com/aconno/sensorics/dagger/configure/ConfigureActivityScope.kt
@@ -4,4 +4,4 @@ import javax.inject.Scope
 
 @Scope
 @Retention(AnnotationRetention.RUNTIME)
-annotation class ConfigurationScope
+annotation class ConfigureActivityScope
