+++ app/src/main/java/com/aconno/sensorics/dagger/deviceselect/DeviceSelectActivityScope.kt
@@ -4,4 +4,4 @@ import javax.inject.Scope
 
 @Scope
 @Retention(AnnotationRetention.RUNTIME)
-annotation class DeviceSelectScope
+annotation class DeviceSelectActivityScope
