+++ app/src/main/java/com/aconno/sensorics/dagger/gcloudpublisher/GoogleCloudPublisherActivityModule.kt
@@ -1,8 +1,6 @@
 package com.aconno.sensorics.dagger.gcloudpublisher
 
 import androidx.lifecycle.ViewModelProviders
-import com.aconno.sensorics.domain.ifttt.GooglePublishRepository
-import com.aconno.sensorics.domain.ifttt.PublishDeviceJoinRepository
 import com.aconno.sensorics.domain.interactor.ifttt.googlepublish.AddGooglePublishUseCase
 import com.aconno.sensorics.domain.interactor.repository.DeletePublishDeviceJoinUseCase
 import com.aconno.sensorics.domain.interactor.repository.SavePublishDeviceJoinUseCase
@@ -14,10 +12,10 @@ import dagger.Module
 import dagger.Provides
 
 @Module
-class GoogleCloudPublisherModule {
+class GoogleCloudPublisherActivityModule {
 
     @Provides
-    @GoogleCloudPublisherScope
+    @GoogleCloudPublisherActivityScope
     fun provideGoogleCloudPublisherViewModel(
         googleCloudPublisherActivity: GoogleCloudPublisherActivity,
         googleCloudPublisherViewModelFactory: GoogleCloudPublisherViewModelFactory
@@ -25,7 +23,7 @@ class GoogleCloudPublisherModule {
         .get(GoogleCloudPublisherViewModel::class.java)
 
     @Provides
-    @GoogleCloudPublisherScope
+    @GoogleCloudPublisherActivityScope
     fun provideGoogleCloudPublisherViewModelFactory(
         addGooglePublishUseCase: AddGooglePublishUseCase,
         googlePublishModelDataMapper: GooglePublishModelDataMapper,
