+++ app/src/main/java/com/aconno/sensorics/dagger/gcloudpublisher/GoogleCloudPublisherActivityScope.kt
@@ -4,4 +4,4 @@ import javax.inject.Scope
 
 @Scope
 @Retention(AnnotationRetention.RUNTIME)
-annotation class GoogleCloudPublisherScope
+annotation class GoogleCloudPublisherActivityScope
