+++ app/src/main/java/com/aconno/sensorics/dagger/mqttpublisher/MqttPublisherActivityScope.kt
@@ -4,4 +4,4 @@ import javax.inject.Scope
 
 @Scope
 @Retention(AnnotationRetention.RUNTIME)
-annotation class MqttPublisherScope
+annotation class MqttPublisherActivityScope
