--- app/src/main/java/com/aconno/sensorics/dagger/publish/PublishScope.kt
@@ -1,7 +1,0 @@
-package com.aconno.sensorics.dagger.publish
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-annotation class PublishScope
