+++ app/src/main/java/com/aconno/sensorics/dagger/publishlist/PublishListActivityModule.kt
@@ -1,0 +1,83 @@
+package com.aconno.sensorics.dagger.publishlist
+
+import androidx.lifecycle.ViewModelProviders
+import com.aconno.sensorics.domain.interactor.ifttt.UpdatePublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.googlepublish.AddGooglePublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.googlepublish.DeleteGooglePublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.googlepublish.GetAllGooglePublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.googlepublish.GetGooglePublishByIdUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.mqttpublish.AddMqttPublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.mqttpublish.DeleteMqttPublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.mqttpublish.GetAllMqttPublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.mqttpublish.GetMqttPublishByIdUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.restpublish.AddRestPublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.restpublish.DeleteRestPublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.restpublish.GetAllRestPublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.restpublish.GetRestPublishByIdUseCase
+import com.aconno.sensorics.model.mapper.*
+import com.aconno.sensorics.ui.settings.publishers.PublishListActivity
+import com.aconno.sensorics.viewmodel.PublishListViewModel
+import com.aconno.sensorics.viewmodel.factory.PublishListViewModelFactory
+import dagger.Module
+import dagger.Provides
+
+/**
+ * @author aconno
+ */
+@Module
+class PublishListActivityModule {
+
+
+    @Provides
+    @PublishListActivityScope
+    fun providePublishListViewModel(
+        activity: PublishListActivity,
+        publishViewModelFactory: PublishListViewModelFactory
+    ) = ViewModelProviders.of(activity, publishViewModelFactory)
+        .get(PublishListViewModel::class.java)
+
+    @Provides
+    @PublishListActivityScope
+    fun providePublishListViewModelFactory(
+        getAllGooglePublishUseCase: GetAllGooglePublishUseCase,
+        getAllRestPublishUseCase: GetAllRestPublishUseCase,
+        googlePublishDataMapper: GooglePublishDataMapper,
+        googlePublishModelDataMapper: GooglePublishModelDataMapper,
+        restPublishDataMapper: RESTPublishDataMapper,
+        restPublishModelDataMapper: RESTPublishModelDataMapper,
+        deleteGooglePublishUseCase: DeleteGooglePublishUseCase,
+        deleteRestPublishUseCase: DeleteRestPublishUseCase,
+        getAllMqttPublishUseCase: GetAllMqttPublishUseCase,
+        mqttPublishModelDataMapper: MqttPublishModelDataMapper,
+        deleteMqttPublishUseCase: DeleteMqttPublishUseCase,
+        updatePublishUseCase: UpdatePublishUseCase,
+        addGooglePublishUseCase: AddGooglePublishUseCase,
+        addRestPublishUseCase: AddRestPublishUseCase,
+        addMqttPublishUseCase: AddMqttPublishUseCase,
+        getGooglePublishByIdUseCase: GetGooglePublishByIdUseCase,
+        getRestPublishByIdUseCase: GetRestPublishByIdUseCase,
+        getMqttPublishByIdUseCase: GetMqttPublishByIdUseCase
+    ) =
+        PublishListViewModelFactory(
+            getAllGooglePublishUseCase,
+            getAllRestPublishUseCase,
+            googlePublishDataMapper,
+            googlePublishModelDataMapper,
+            restPublishDataMapper,
+            restPublishModelDataMapper,
+            deleteGooglePublishUseCase,
+            deleteRestPublishUseCase,
+            getAllMqttPublishUseCase,
+            mqttPublishModelDataMapper,
+            deleteMqttPublishUseCase,
+            updatePublishUseCase,
+            addGooglePublishUseCase,
+            addRestPublishUseCase,
+            addMqttPublishUseCase,
+            getGooglePublishByIdUseCase,
+            getRestPublishByIdUseCase,
+            getMqttPublishByIdUseCase
+        )
+
+
+}
