+++ app/src/main/java/com/aconno/sensorics/dagger/publishlist/PublishListActivityScope.kt
@@ -1,4 +1,4 @@
-package com.aconno.sensorics.dagger.actionlist
+package com.aconno.sensorics.dagger.publishlist
 
 import javax.inject.Scope
 
@@ -7,4 +7,4 @@ import javax.inject.Scope
  */
 @Scope
 @Retention(AnnotationRetention.RUNTIME)
-annotation class ActionListScope
+annotation class PublishListActivityScope
