+++ app/src/main/java/com/aconno/sensorics/dagger/restpublisher/RESTPublisherActivityScope.kt
@@ -4,4 +4,4 @@ import javax.inject.Scope
 
 @Scope
 @Retention(AnnotationRetention.RUNTIME)
-annotation class RESTPublisherScope
+annotation class RESTPublisherActivityScope
